Add modest error checking and modify struct
Ensure an input key corresponds to a value before sending it to the
client. Likewise, ensure a key-value entry was in fact added to the hash
table.

Rename "bundle" struct to "unit" and split its "entry" member into a
separate key member and a separate value member because they are passed
into dict_add() individually regardless.